+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * PPU issue related to the maximum number of objects/sprite per line being 10, issue detected by ACID test
 * Object pixel drawing priority issue, issue detected by ACID test
 * Issue associated with the wrongful flipping of 8x16 sprites, issue detected by ACID test
+* Issue associated with drawing of window tiles, due to extra `update_stat()` operations, issue detected by ACID test
 
 ## [0.5.4] - 2022-11-15